+++ src/c_cmds.c
@@ -736,6 +736,8 @@ consolecmd_t consolecmds[] =
         "Your armor type (" BOLD("none") ", " BOLD("green") " or " BOLD("blue") ")."),
     BOOLCVAR(autoaim, "", "", boolfunc1, boolfunc2, 0,
         "Toggles vertical autoaiming as you fire your weapon while using freelook."),
+    BOOLCVAR(autofire, "", "", boolfunc1, boolfunc2, 0,
+        "Toggles automatically firing your currently equipped weapon if there is a monster in front of you."),
     BOOLCVAR(autoload, "", "", boolfunc1, boolfunc2, CF_PISTOLSTART,
         "Toggles automatically loading the last savegame when you die."),
     BOOLCVAR(autosave, "", "", boolfunc1, boolfunc2, 0,